pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["YuZhanHuang <bigjoe1992111@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "auth_service"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "2.2.2"
flask-sqlalchemy = "3.0.2"
flask-redis = "0.4.0"
flask-migrate = "3.1.0"
gunicorn = "20.1.0"
structlog = "23.1.0"
voluptuous = "0.13.1"
flask-security = "3.0.0"
flask-bcrypt = "1.0.1"
flask-cors = "3.0.10"
flask-jwt-extended = "4.4.4"
pydantic = "1.10.7"
delorean = "1.0.0"
psycopg2-binary = "2.9.6"
gevent = "22.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
